s3/lib: Fix CID 1659589 - protect from underflow
Refine code to hint Coverity scan that there is no real overflow when
converting from signed to unsigned.

+++ source3/lib/recvfile.c
@@ -298,7 +298,7 @@ ssize_t drain_socket(int sockfd, size_t count)
 			count = (size_t)-1;
 			goto out;
 		}
-		total += read_ret;
+		total += (size_t)read_ret;
 	}
 
   out:
